# bondly-api/docs/swagger.yaml
basePath: /
definitions:
...
  models.Transaction:
    properties:
      block_number:
        type: integer
      created_at:
        type: string
      from_address:
        type: string
      gas_price:
        type: string
      gas_used:
        type: integer
      hash:
        type: string
      id:
        type: integer
      status:
        description: pending, confirmed, failed
        type: string
      to_address:
        type: string
      updated_at:
        type: string
      value:
        type: string
    type: object
...
  response.ResponseAny:
    properties:
      code:
        example: 1000
        type: integer
      data: {}
      message:
        example: 操作成功
        type: string
      success:
        example: true
        type: boolean
    type: object
...
  response.ResponseTransaction:
    properties:
      code:
        example: 1000
        type: integer
      data: {}
      message:
        example: 操作成功
        type: string
      success:
        example: true
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@bondly.com
    name: Bondly API Support
    url: https://bondly.com/support
  description: 这是Bondly项目的API文档，提供用户认证、区块链交互、内容管理和治理功能。
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://bondly.com/terms
  title: Bondly API
  version: "1.0"
paths:
...
  /api/v1/transactions:
    get:
      consumes:
      - application/json
      description: 分页获取交易列表
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseTransaction'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ResponseAny'
      summary: 获取交易列表
      tags:
      - 交易管理
    post:
      consumes:
      - application/json
      description: 创建新的交易记录
      parameters:
      - description: 交易信息
        in: body
        name: transaction
        required: true
        schema:
          $ref: '#/definitions/models.Transaction'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.ResponseTransaction'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ResponseAny'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ResponseAny'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ResponseAny'
      security:
      - BearerAuth: []
      summary: 创建交易
      tags:
      - 交易管理
  /api/v1/transactions/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定的交易
      parameters:
      - description: 交易ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseAny'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ResponseAny'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ResponseAny'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ResponseAny'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ResponseAny'
      security:
      - BearerAuth: []
      summary: 删除交易
      tags:
      - 交易管理
    get:
      consumes:
      - application/json
      description: 根据ID获取交易详情
      parameters:
      - description: 交易ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseTransaction'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ResponseAny'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ResponseAny'
      summary: 获取交易详情
      tags:
      - transaction
    put:
      consumes:
      - application/json
      description: 更新指定交易的信息
      parameters:
      - description: 交易ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新的交易信息
        in: body
        name: transaction
        required: true
        schema:
          $ref: '#/definitions/models.Transaction'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseTransaction'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ResponseAny'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ResponseAny'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ResponseAny'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ResponseAny'
      security:
      - BearerAuth: []
      summary: 更新交易
      tags:
      - 交易管理
  /api/v1/transactions/hash/{hash}:
    get:
      consumes:
      - application/json
      description: 根据交易哈希获取交易详情
      parameters:
      - description: 交易哈希
        in: path
        name: hash
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseTransaction'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ResponseAny'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ResponseAny'
      summary: 根据哈希获取交易
      tags:
      - 交易管理
  /api/v1/transactions/stats:
    get:
      consumes:
      - application/json
      description: 获取交易相关的统计数据
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseAny'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ResponseAny'
      summary: 获取交易统计
      tags:
      - 交易管理
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"